[OF-1772] fix: resolve unsupported ReplicatedValue type in ai chatbot component

+++ Library/src/Multiplayer/Components/AIChatbotComponent.cpp
@@ -27,7 +27,7 @@ csp::multiplayer::AIChatbotSpaceComponent::AIChatbotSpaceComponent(csp::common::
     Properties[static_cast<uint32_t>(AIChatbotPropertyKeys::Position)] = csp::common::Vector3::Zero();
     Properties[static_cast<uint32_t>(AIChatbotPropertyKeys::Voice)] = "";
     Properties[static_cast<uint32_t>(AIChatbotPropertyKeys::GuardrailAssetCollectionId)] = "";
-    Properties[static_cast<uint32_t>(AIChatbotPropertyKeys::VisualState)] = static_cast<uint32_t>(0);
+    Properties[static_cast<uint32_t>(AIChatbotPropertyKeys::VisualState)] = static_cast<int64_t>(0);
 
     SetScriptInterface(new AIChatbotSpaceComponentScriptInterface(this));
 }
